fourth_project/requirements.txt
django
python-dotenv
markdown2

numpy
Pillow
PyMuPDF

torch
torchvision
sentence-transformers
langchain
langchain-community
langchain-openai
faiss-cpu
paddleocr
paddlepaddle

setuptools
wheel